\name{checkDims}
\alias{checkDims}
\title{Check the consistency of the dimensions of a LP or MIP}
\usage{
checkDims(p)
}
\arguments{
  \item{p}{an object of class \code{lpmodeler}}
}
\description{
\code{checkDims} checks the consistency of the dimensions
of the matrices and vectors of a linear program (LP) or a
mixed integer program (MIP) reprensented by an object of
class \code{lpmodeler}.
}
\author{
Cyrille Szymanski <cnszym at gmail.com>
}
\seealso{
TODO
}

