\name{lpmodeler-package}
\alias{lpmodeler}
\docType{package}
\title{
Modeler for linear programs (LP) and mixed integer programs (MIP)
}
\description{
\code{lpmodeler} is a set of user-friendly functions to simplify the
modelling of linear programs (LP) and mixed integer programs (MIP).
It provides a unified interface compatible with optimization packages: 
Rsymphony.
}
\details{
TODO
}
\author{
Cyrille Szymanski <cnszym@gmail.com>
}
\references{
TODO: Papers, books
}
\keyword{ package }
\seealso{
% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
TODO: R packages
TODO: other software
}
\examples{
# TODO
}
